+++ internal/packager/packager_test.go
@@ -7,14 +7,14 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestInterfaceSatisfaction(t *testing.T) {
+func TestAccInterfaceSatisfaction(t *testing.T) {
 	t.Parallel()
 
 	var _ ZIPI = &ZIP{}
 	var _ ZIPI = &MockZIP{}
 }
 
-func TestZIPZip(t *testing.T) {
+func TestAccZIPZip(t *testing.T) {
 	t.Parallel()
 
 	t.Cleanup(func() {
@@ -22,7 +22,7 @@ func TestZIPZip(t *testing.T) {
 	})
 
 	zip := ZIP{}
-	hash, err := zip.Zip("test.zip", map[string]string{
+	err := zip.Zip("test.zip", map[string]string{
 		"packager.go":      "packager.go",
 		"packager_mock.go": "a/packager_mock.go",
 		"packager_test.go": "b/packager_test.go",
@@ -30,5 +30,4 @@ func TestZIPZip(t *testing.T) {
 	})
 
 	assert.NoError(t, err)
-	assert.Regexp(t, `^[0-9a-zA-Z]+$`, hash)
 }